Rename from TransformSync to VariableSync
